Update MainActivity.kt
What This Code Does:

WebView: This sets up the WebView to display a website (in this case, Google).

JavaScript: It enables JavaScript in case the webpage requires it to function properly.

WebViewClient: Ensures that links clicked within the WebView stay inside the app and don't open in a separate browser.