Create activity_main.xml
Explanation:
LinearLayout: This layout will organize the child elements (in this case, the WebView) vertically (top to bottom), as the orientation="vertical" is set.
WebView: A UI component to display web content.
android:id="@+id/myWebView": This assigns an ID to the WebView, which will be used to reference it in your Kotlin code.
android:layout_width="match_parent" and android:layout_height="match_parent": These properties make the WebView take up the full width and height of the parent layout.